app/requirements.txt
black==23.10.0
Django==4.2.6
djangorestframework==3.14.0
drf-yasg==1.21.7
flake8==6.1.0
isort==5.12.0
psycopg2-binary==2.9.9
pytest==7.4.2
pytest-django==4.5.2
requests==2.31.0